refactor(adr-061): split Itanium mangled-name parser into model, classify dumper_clang_expr extract
Same root-cause pattern as the binary_naming.py and cc_attributes.py
splits, at larger scale: diff_cxx_rules.py's itanium_scope_components()
and its private parsing chain (~380 lines: _itanium_strip_prefix,
_parse_source_name_component, _parse_ctor_dtor_component,
_parse_operator_component, _step_next_component,
itanium_scope_components_with_template_positions, plus the
_ASCII_DIGITS/_ITANIUM_OPERATORS constants) is pure string decoding with
no I/O and no dependency on model's own entity types -- confirmed before
moving anything: the only ChangeKind/Change/Function/RecordType mention
in that range was a single docstring cross-reference.

Two extract modules already worked around not being able to import it
cleanly (dumper_clang_expr.py carried an explicit comment about the
compare-classified dependency it couldn't take; dumper_hybrid.py just
imported it anyway, an existing, real extract -> compare inversion).
compare's own diff_cxx_rules.py needs the same chain internally
(owner_class_of's mangled-name fallback, itanium_ctor_dtor_marker_span).

Moved into model/mangled_name.py. diff_cxx_rules.py re-exports every
name it still uses by value for back-compat; dumper_clang_expr.py and
dumper_hybrid.py now import the canonical path directly, closing the
workaround comment and the inversion respectively. dumper_clang_expr.py
now classifies extract in architecture/modules.yaml. msvc_scope_components
and the real vtable/virtual-method diff logic (owner_class_of onward)
stay in diff_cxx_rules.py -- genuinely compare-shaped.

Verification: check_architecture.py 0 errors; mypy/ruff clean; 2936
targeted tests (diff_types/diff_symbols/dumper/mangled-name family) plus
the architecture suite (119) all pass; new tests/test_mangled_name.py
pins the split and both re-export identities.
