+++ changelog.d/1788019943_claude_adr061_mangled_name_split.md
@@ -1,0 +1,3 @@
+### Changed
+
+- **Itanium mangled-name scope parsing split into `model/mangled_name.py`, classifying `dumper_clang_expr.py` `extract` under ADR-061** — the biggest of this round's root-cause splits: `diff_cxx_rules.py`'s `itanium_scope_components()` and its ~380-line private parsing chain (`_itanium_strip_prefix`, `_parse_source_name_component`, `_parse_ctor_dtor_component`, `_parse_operator_component`, `_step_next_component`, `itanium_scope_components_with_template_positions`, plus the `_ASCII_DIGITS`/`_ITANIUM_OPERATORS` constants) is pure string decoding with no I/O and no dependency on `model`'s own entity types — verified by grep before moving anything: the only `ChangeKind`/`Change`/`Function`/`RecordType` mention anywhere in that range was a single docstring cross-reference. It was needed by two `extract`-classified modules already reaching around the direction rule with an inline workaround comment (`dumper_clang_expr.py`, `dumper_hybrid.py`), by `compare`'s own `diff_cxx_rules.py` (`owner_class_of`'s mangled-name fallback, `itanium_ctor_dtor_marker_span`), and transitively by several other `compare`/`policy` modules via `owner_class_of`/`itanium_qualified_name` — exactly the shared-leaf shape `qualified_name_segments.py`/`binary_naming.py`/`cc_attributes.py` already established. Moved into `model/mangled_name.py`; `diff_cxx_rules.py` re-exports every name it still uses internally by value for back-compat, `dumper_clang_expr.py` and `dumper_hybrid.py` now import the canonical path directly (closing out the workaround comment `dumper_clang_expr.py` had carried since an earlier ADR-061 phase), and `dumper_clang_expr.py` classifies cleanly as `extract` in `architecture/modules.yaml`. `msvc_scope_components`, `_resolve_owner_type`, `virtual_signature_key`, and the rest of `diff_cxx_rules.py`'s real vtable/virtual-method diff logic stay put — genuinely `compare`-shaped, not name parsing. No behavior change — same function objects either way, pinned by `tests/test_mangled_name.py`; the full existing `itanium_scope_components` behavioral coverage (`test_dumper_hybrid.py`, `test_dumper_clang.py`, `test_type_reachability_mangling.py`, and others) still passes unmodified against the moved code.
